Splitting Plasma Contract into Base and Auction functionality
